Former-commit-id: a14c7dba86bc31780a8589619b057b03a48a7915

+++ framework/sofa/helper/random.h
@@ -51,7 +51,7 @@ inline int irand()
 // Random double between 0 and 1 [0.0,1.0]
 inline double drand()
 {
-   return std::rand() / RAND_MAX;
+   return double(std::rand()) / RAND_MAX;
 }
 
 // Random double between 0 and max [0.0,max]
@@ -59,7 +59,6 @@ inline double drand()
 inline double drandpos(double max)
 {
    assert( max >= 0 );
-
    return drand() * max;
 }
